Fix bug in DoubleFeatureValue.
Value set by #set(Double) was the intValue instead of doubleValue.

+++ src/main/java/org/mastodon/features/DoubleFeatureValue.java
@@ -33,7 +33,7 @@ public final class DoubleFeatureValue< O > implements FeatureValue< Double >
 		if ( value == null )
 			featureMap.remove( object );
 		else
-			featureMap.put( object, value.intValue() );
+			featureMap.put( object, value.doubleValue() );
 	}
 
 	public void set( final double value )
